# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manga Collection API",
        "description": "Manga Collection API Docs",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "local server"
        },
        {
            "url": "https://pink-buffalo-749567.hostingersite.com/api",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Index",
                "operationId": "0da298f7b8d268eaed86946671c66451",
                "parameters": [
                    {
                        "name": "ordering",
                        "in": "query",
                        "description": "Parameter to order results",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Parameter that set the page",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Parameter that set the size of the result collection",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of genres"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Store",
                "operationId": "9bd20b05d63f461c6457722421e36470",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Action"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Genre created successfully"
                    },
                    "422": {
                        "description": "Field Error"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "499": {
                        "description": "Not allowed"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/genres/{genre_id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Show",
                "operationId": "105fe422173c8ed8ddf05b3d5ca4a52d",
                "parameters": [
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the requested genre"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Destroy",
                "operationId": "2539c871e81a94085c85db0f0494893a",
                "parameters": [
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Genre deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Genres"
                ],
                "summary": "Update",
                "operationId": "f7d854b7c1c96b1f0fb9216f9ecb63b9",
                "parameters": [
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "Parameter that filter the entity",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Genre updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Not allowed"
                    },
                    "499": {
                        "description": "Not allowed"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}